+++ src/main/java/de/bixilon/minosoft/data/mappings/blocks/actions/MobSpawnerAction.java
@@ -15,7 +15,7 @@ package de.bixilon.minosoft.data.mappings.blocks.actions;
 
 public class MobSpawnerAction implements BlockAction {
 
-    public MobSpawnerAction(byte status, byte ignored) {
+    public MobSpawnerAction(short status, short ignored) {
         // only 1 action (id 1)
     }
